Rename folder common
